module-02/requirements.txt
langchain
langchain-core
langchain-ollama
python-dotenv
requests
pydantic
